refactor(analysis): refactor dfd analysis into seperate plugin
BREAKING CHANGE: Namespace changed for dfd components
