+++ bundles/org.dataflowanalysis.analysis.dfd/META-INF/MANIFEST.MF
@@ -1,0 +1,18 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: DataFlowAnalysis DFD
+Bundle-SymbolicName: org.dataflowanalysis.analysis.dfd
+Bundle-Version: 1.0.0.qualifier
+Require-Bundle: org.dataflowanalysis.analysis;bundle-version="1.0.0",
+ org.eclipse.core.runtime;bundle-version="3.24.0",
+ org.eclipse.core.resources;bundle-version="3.16.0";visibility:=reexport,
+ tools.mdsd.library.standalone.initialization;bundle-version="0.2.0";visibility:=reexport,
+ tools.mdsd.library.standalone.initialization.log4j;bundle-version="0.2.0";visibility:=reexport,
+ org.apache.log4j;bundle-version="1.2.15";visibility:=reexport,
+ org.dataflowanalysis.dfd.dataflowdiagram;bundle-version="0.1.0"
+Bundle-RequiredExecutionEnvironment: JavaSE-17
+Automatic-Module-Name: org.dataflowanalysis.analysis.dfd
+Export-Package: org.dataflowanalysis.analysis.dfd.builder,
+ org.dataflowanalysis.analysis.dfd.core,
+ org.dataflowanalysis.analysis.dfd,
+ org.dataflowanalysis.analysis.dfd.resource
