+++ bundles/org.dataflowanalysis.analysis.dfd/build.properties
@@ -1,0 +1,3 @@
+source.. = src/
+bin.includes = META-INF/,\
+               .
