+++ bundles/org.dataflowanalysis.analysis.dfd/src/org/dataflowanalysis/analysis/dfd/builder/DFDAnalysisBuilderData.java
@@ -1,9 +1,14 @@
-package org.dataflowanalysis.analysis.builder.dfd;
+package org.dataflowanalysis.analysis.dfd.builder;
+
+import java.util.ArrayList;
+import java.util.List;
 
 import org.dataflowanalysis.analysis.AnalysisData;
 import org.dataflowanalysis.analysis.builder.AnalysisBuilderData;
-import org.dataflowanalysis.analysis.resource.dfd.DFDResourceProvider;
-import org.dataflowanalysis.analysis.resource.dfd.DFDURIResourceProvider;
+import org.dataflowanalysis.analysis.builder.validation.ValidationError;
+import org.dataflowanalysis.analysis.builder.validation.ValidationErrorSeverity;
+import org.dataflowanalysis.analysis.dfd.resource.DFDResourceProvider;
+import org.dataflowanalysis.analysis.dfd.resource.DFDURIResourceProvider;
 import org.dataflowanalysis.analysis.utils.ResourceUtils;
 
 public class DFDAnalysisBuilderData extends AnalysisBuilderData {
@@ -12,13 +17,15 @@ public class DFDAnalysisBuilderData extends AnalysisBuilderData {
 	
 	public DFDAnalysisBuilderData() {}
 	
-	public void validateData() {
+	public List<ValidationError> validate() {
+		List<ValidationError> validationErrors = new ArrayList<>(super.validate());
 		if (this.dataDictionaryPath == null || this.dataDictionaryPath.isBlank()) {
-			throw new IllegalStateException("A path to a data dictionary is required");
+			validationErrors.add(new ValidationError("A path to a data dictionary is required", ValidationErrorSeverity.ERROR));
 		}
 		if (this.dataFlowDiagramPath == null || this.dataFlowDiagramPath.isBlank()) {
-			throw new IllegalStateException("A path to a data flow diagram is required");
+			validationErrors.add(new ValidationError("A path to a data flow diagram is required", ValidationErrorSeverity.ERROR));
 		}
+		return validationErrors;
 	}
 	
 	public void setDataFlowDiagramPath(String dataFlowDiagramPath) {
