+++ bundles/org.dataflowanalysis.analysis.dfd/src/org/dataflowanalysis/analysis/dfd/builder/DFDDataFlowAnalysisBuilder.java
@@ -1,12 +1,17 @@
-package org.dataflowanalysis.analysis.builder.dfd;
+package org.dataflowanalysis.analysis.dfd.builder;
 
+import java.util.List;
 import java.util.Optional;
 
-import org.dataflowanalysis.analysis.DFDConfidentialityAnalysis;
+import org.apache.log4j.Logger;
 import org.dataflowanalysis.analysis.builder.DataFlowAnalysisBuilder;
+import org.dataflowanalysis.analysis.builder.validation.ValidationError;
+import org.dataflowanalysis.analysis.dfd.DFDConfidentialityAnalysis;
 import org.eclipse.core.runtime.Plugin;
 
 public class DFDDataFlowAnalysisBuilder extends DataFlowAnalysisBuilder {
+	private final Logger logger = Logger.getLogger(DFDDataFlowAnalysisBuilder.class);
+	
 	private DFDAnalysisBuilderData builderData;
 
 	public DFDDataFlowAnalysisBuilder() {
@@ -59,8 +64,16 @@ public class DFDDataFlowAnalysisBuilder extends DataFlowAnalysisBuilder {
 		return this;
 	}
 
+	/**
+	 * Builds a new analysis from the given data
+	 */
 	public DFDConfidentialityAnalysis build() {
-		this.builderData.validateData();
+		List<ValidationError> validationErrors = this.builderData.validate();
+		validationErrors.forEach(it -> it.log(logger));
+		if (validationErrors.stream().anyMatch(ValidationError::isFatal)) {
+			logger.fatal("Could not create analysis due to fatal validation errors");
+			throw new IllegalStateException();
+		}
 		return new DFDConfidentialityAnalysis(this.builderData.createAnalysisData(), this.builderData.getPluginActivator(), this.builderData.getModelProjectName());
 	}
 }
