+++ bundles/org.dataflowanalysis.analysis/META-INF/MANIFEST.MF
@@ -18,9 +18,7 @@ Require-Bundle: org.eclipse.core.runtime;bundle-version="3.24.0",
 Export-Package: 
  org.dataflowanalysis.analysis,
  org.dataflowanalysis.analysis.builder,
- org.dataflowanalysis.analysis.builder.dfd,
+ org.dataflowanalysis.analysis.builder.validation,
  org.dataflowanalysis.analysis.core,
- org.dataflowanalysis.analysis.core.dfd,
  org.dataflowanalysis.analysis.resource,
- org.dataflowanalysis.analysis.utils,
- org.dataflowanalysis.analysis.builder.validation
+ org.dataflowanalysis.analysis.utils
