+++ features/org.dataflowanalysis.analysis.feature/feature.xml
@@ -21,4 +21,11 @@
          version="0.0.0"
          unpack="false"/>
 
+   <plugin
+         id="org.dataflowanalysis.analysis.dfd"
+         download-size="0"
+         install-size="0"
+         version="0.0.0"
+         unpack="false"/>
+
 </feature>
