+++ tests/org.dataflowanalysis.analysis.tests/META-INF/MANIFEST.MF
@@ -14,4 +14,5 @@ Require-Bundle: org.dataflowanalysis.analysis;bundle-version="1.0.0",
  org.dataflowanalysis.pcm.extension.nodecharacteristics;bundle-version="0.1.0",
  org.dataflowanalysis.dfd.datadictionary;bundle-version="0.1.0",
  org.dataflowanalysis.dfd.dataflowdiagram;bundle-version="0.1.0",
- org.dataflowanalysis.analysis.pcm;bundle-version="1.0.0"
+ org.dataflowanalysis.analysis.pcm;bundle-version="1.0.0",
+ org.dataflowanalysis.analysis.dfd;bundle-version="1.0.0"
